+++ src/pages/Dashboard/Dashboard.css
@@ -7,7 +7,6 @@ body {
 }
 
 .background-dashboard {
-  background-color: #1f2a37;
   display: flex;
   flex-direction: column;
   min-height: 100vh;
